requirement.txt
elasticsearch==8.12.0
python-dotenv==1.0.0
sentence_transformers==2.7.0
ibm-watsonx-ai==1.0.4